modification background menu et suppression focntions inutiles

+++ Projet/src/init.h
@@ -16,7 +16,7 @@ int notre_init(int argc, char** argv, void (*Modelisation)());
 void initialise();
 void reinitialise();
 
-int Rand(int a, int b);
+float Rand(float a, float b);
 void remplissageTableauR();
 void creationAsteroides();
